HTML/CSS Redo
realized i was overcomplicating things for myself in terms of trying to use scss & bootstrap so i decided to keep things simple. site looks a little worse but honestly that's kinda the goal with the site. (also this is my only coding experience so it's par for the course lmao)